Commit SW 2023

+++ app/src/main/java/com/example/spendwisely/data/repositorios/CuentaRepository.kt
@@ -3,12 +3,13 @@ package com.example.spendwisely.data.repositorios
 import androidx.lifecycle.LiveData
 import com.example.spendwisely.data.daos.CuentaDao
 import com.example.spendwisely.data.entidades.Cuenta
+import java.math.BigDecimal
 
 class CuentaRepository(private val cuentaDao: CuentaDao) {
 
     val allCuentas: LiveData<List<Cuenta>> = cuentaDao.getAllCuentas()
 
-    val sumSaldoCuentas : LiveData<Double> = cuentaDao.sumSaldoCuentas()
+    val sumSaldoCuentas : LiveData<BigDecimal> = cuentaDao.sumSaldoCuentas()
 
     suspend fun insert(cuenta: Cuenta) {
         cuentaDao.addCuenta(cuenta)
@@ -22,11 +23,11 @@ class CuentaRepository(private val cuentaDao: CuentaDao) {
         cuentaDao.deleteCuenta(cuenta)
     }
 
-    suspend fun restarSaldo(gasto: Double) {
+    suspend fun restarSaldo(gasto: BigDecimal) {
         cuentaDao.restarSaldo(gasto)
     }
 
-    suspend fun sumarSaldo(saldo: Double) {
+    suspend fun sumarSaldo(saldo: BigDecimal) {
         cuentaDao.sumarSaldo(saldo)
     }
